Move config utils to llm.config
