+++ tests/config_test.py
@@ -1,10 +1,14 @@
 from __future__ import annotations
 
 import pathlib
+from typing import Any
+from unittest import mock
 
 import pytest
 
 from llm.config import Config
+from llm.config import flatten_mapping
+from llm.config import flattened_config
 from llm.config import load_config
 
 
@@ -31,6 +35,44 @@ def test_config_nested_dict() -> None:
     config = Config(**d)
 
     assert config.a.b == 2
+
+
+def test_flattened_config() -> None:
+    assert flattened_config({'a': 1}) == {'a': 1}
+
+    config = Config(**{'a': 1})
+    assert flattened_config(config) == {'a': 1}
+
+    # test get world size
+    with mock.patch(
+        'torch.distributed.is_initialized',
+        return_value=True,
+    ), mock.patch('torch.distributed.get_world_size', return_value=4):
+        assert flattened_config() == {'world_size': 4}
+
+    # test flattening and removing non-supported types
+    config = Config(**{'a': {'b': 1, 'c': [1, 2]}})
+    assert flattened_config(config) == {'a_b': 1}
+
+
+@pytest.mark.parametrize(
+    'in_,out,kwargs',
+    (
+        ({}, {}, {}),
+        ({'a': 1}, {'a': 1}, {}),
+        ({'a': {'b': 1}}, {'a_b': 1}, {}),
+        ({'a': {'b': 1}}, {'a-b': 1}, {'sep': '-'}),
+        ({'a': {'b': 1}}, {'x_a_b': 1}, {'parent': 'x'}),
+        ({'a': {'b': {'c': 1}}}, {'a_b_c': 1}, {}),
+        ({'a': {'b': 1}, 'c': 2}, {'a_b': 1, 'c': 2}, {}),
+    ),
+)
+def test_flatten_mapping(
+    in_: dict[str, Any],
+    out: dict[str, Any],
+    kwargs: dict[str, Any],
+) -> None:
+    assert flatten_mapping(in_, **kwargs) == out
 
 
 def test_load_config(tmp_path: pathlib.Path) -> None:
